refactor: move the utility fn to a file
