+++ packages/analytics-js-common/src/utilities/detect.ts
@@ -1,0 +1,8 @@
+/**
+ * Determines if the SDK is running inside a chrome extension
+ * @returns boolean
+ */
+const isSDKRunningInChromeExtension = (): boolean =>
+  !!((window as any).chrome && (window as any).chrome.runtime && (window as any).chrome.runtime.id);
+
+export { isSDKRunningInChromeExtension };
