+++ packages/analytics-js-common/src/utilities/index.ts
@@ -11,3 +11,4 @@ export * as timestamp from './timestamp';
 export * as url from './url';
 export * as uuId from './uuId';
 export * as timezone from './timezone';
+export * as detect from './detect';
